make Word component independent
